added create parser link feature endpoint

+++ src/main_app/RemTech.Solution/modules/scrapers/Scrapers.Module/Features/CreateNewParser/Cache/CachedParserLink.cs
@@ -6,7 +6,7 @@ internal sealed record CachedParserLink(
     string Url,
     bool Activity,
     int Processed,
-    int TotalSeconds,
+    long TotalSeconds,
     int Hours,
     int Minutes,
     int Seconds
